+++ src/features/index.ts
@@ -4,6 +4,7 @@ import { register as about } from "./about";
 import { register as checklistProgress } from "./checklistProgress";
 import { register as checklistToggle } from "./checklistToggle";
 import { register as citationGenerator } from "./citationGenerator";
+import { register as cleanPaste } from "./cleanPaste";
 import { register as dailyNote } from "./dailyNote";
 import { register as newReference } from "./newReference";
 import { register as wordCount } from "./wordCount";
@@ -13,6 +14,7 @@ export function register(context: vscode.ExtensionContext) {
   checklistProgress(context);
   checklistToggle(context);
   citationGenerator(context);
+  cleanPaste(context);
   dailyNote(context);
   newReference(context);
   wordCount(context);